% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-to-html.R
\name{md_to_html}
\alias{md_to_html}
\title{Convert markdown to html}
\usage{
md_to_html(text)
}
\arguments{
\item{text}{Character string}
}
\description{
Convert a markdown string into an html string.
}
\references{
https://commonmark.org/
}
