% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-markdown.R
\name{render_markdown}
\alias{render_markdown}
\title{Render HTML or Markdown to a list of grid grobs}
\usage{
render_markdown(text, css = "", width_in = 5, height_in = 5)
}
\arguments{
\item{text}{HTML or Markdown to render}

\item{css}{The CSS to use when rendering `text`}

\item{width_in}{Width of the drawing area, in inches}

\item{height_in}{Height of the drawing area, in inches}
}
\description{
Render HTML or Markdown to a list of grid grobs.
}
