% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_grob.R
\name{html_grob}
\alias{html_grob}
\title{Render HTML or Markdown}
\usage{
html_grob(
  text,
  x = unit(0.1, "npc"),
  y = unit(0.9, "npc"),
  width = unit(0.8, "npc"),
  height = unit(0.8, "npc"),
  hjust = 0,
  vjust = 1,
  css = "",
  vp = NULL
)
}
\arguments{
\item{text}{HTML or Markdown text to render}

\item{x, y}{x and y position}

\item{width, height}{width and height}

\item{hjust, vjust}{horizontal and vertical justification relative
to `x` and `y`}

\item{css}{CSS specification to use for rendering}

\item{vp}{viewport}
}
\description{
Render HTML or Markdown
}
