% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-to-html.R
\name{md_to_html}
\alias{md_to_html}
\title{Convert markdown to html}
\usage{
md_to_html(text)
}
\arguments{
\item{text}{Character vector. If of length > 1, each string is
separately converted.}
}
\value{
Character vector of html fragments.
}
\description{
Convert a markdown string into an html string. The markdown conversion
adheres to the commonmark standard with tables extension.
}
\references{
https://commonmark.org/
}
